Ensure temp relation files are removed from the file-system on a DROP.
This fixes issue:#7922. For a DROP on a relation, after the COMMIT
PREPARED record is written on the QE, we remove the relation file from
disk (call to DropRelationFiles() inside FinishPreparedTransaction()).
While doing so, we need to know whether the relation is a temp relation
such that the file deletion code can prepend the 't_' prefix to the
relfilenode name.  Apart from the case above, whenever we have code
dropping temporary relation files as a part of the 2PC mechanism, we
have to ensure that we have enough information to construct the correct
filesystem path to the temp relation's relfilenode.

This commit ensures that we persist a flag specifying whether a relation
is temporary through the 2PC infrastructure (a variety of XLOG records
and in-mem pending deletes structure), in order to access it while
performing the file removal inside DropRelationFiles.