Introduce catalog function to print GP_TABLESPACE_VERSION_DIRECTORY
GP_TABLESPACE_VERSION_DIRECTORY's value varies with the catalog version
and is something tablespace related tests rely on. Thus we introduce
gp_tablespace_version_directory_name(). Further, this function exposes
this constant to external utilities where we anticipate its need.

Previously we had the non-catalog function equivalent called
get_tablespace_version_directory_name() which was part of the regress
framework. This commit tears it out and replaces it with the new
function.

Co-authored-by: Taylor Vesely <tvesely@pivotal.io>