6X backport: Add support for update extension
This backports commit: 841d8fa6946

The implementation of UPDATE EXTENSION is very similar to CREATE
EXTENSION.

First, on QD, we update the pg_extension row to the new version, and
update extension's dependency entries according
to the new requires introduced by the new version specified control
file.

Second, we dispatch the UPDATE EXTENSION statement to QEs to update
pg_extension row and extension's dependency entries on QEs, and the
global status like creating_extension, CurrentExtensionObject of QEs
will be set.

Finally, we execute the update script file on QD. because of the correct
set of creating_extension, CurrentExtensionObject on QEs, the dependency
between object and extension will be handled correctly, so the
CREATE/DROP object statement will be executed successfully.

Fixes #8862 github issue.

Co-authored-by: Soumyadeep Chakraborty <sochakraborty@pivotal.io>