WIP: Initial changes to vacuum flow
So far, takes dispatch out of `vacuum_rel`.

But still getting this:
```
FATAL:  Unexpected internal error (xact.c:2545)
DETAIL:  FailedAssertion("!(s->prevSecContext == 0)", File: "xact.c", Line: 2545)
server closed the connection unexpectedly
	This probably means the server terminated abnormally
	before or while processing the request.
The connection to the server was lost. Attempting reset: Failed.
```

This is because we are setting the security context outside Tx
boundaries. We have to start the Tx before the dispatch, like in 5X.

Co-authored-by: Soumyadeep Chakraborty <sochakraborty@pivotal.io>