Properly mark null return from combine functions
We had a bug in a few of the combine functions where if the combine
function returned a NULL, it didn't set fcinfo->isnull = true. This led
to a segfault when we would spill in the final hashagg of a two-stage
agg inside the serial function. So, properly mark NULL outputs from the
combine functions.

Co-authored-by: Denis Smirnov <sd@arenadata.io>
Co-authored-by: Soumyadeep Chakraborty <sochakraborty@pivotal.io>