pg_rewind test suite: Bump the wait time for standby streaming/promotion
This follows on from commits 9c66b2b106e, 07435e06629 and 5328564aeea,
which all include changes for fixing the wait time for
streaming/promotion to complete.

These changes are in light of recent timeouts (especially for streaming)
in CI.

This commit conforms the wait and retries to that of the corresponding
TAP tests for pg_rewind. Refer poll_query_until.