check_gp.c: Use pg_catalog to empty out search path
Setting search_path to '' in 5X, during a 5X->6X upgrade leads to:

SET search_path to '';
ERROR:  schema "" does not exist.

So, set search_path to 'pg_catalog' instead, which has the same effect.

PG Documentation excerpt: The system catalog schema, pg_catalog, is
always searched, whether it is mentioned in the path or not. If it is
mentioned in the path then it will be searched in the specified order.
If pg_catalog is not in the path then it will be searched before
searching any of the path items.

Co-authored-by: David Krieger <dkrieger@vmware.com>
Co-authored-by: Kevin Yeap <kyeap@vmware.com>